Tests unitaires

+++ BibliothequeSkill/src/fr/ynov/biblioskill/service/Dao.java
@@ -4,9 +4,9 @@ import java.util.List;
 
 public interface Dao<T> {
 	public List<T> lireTous();
-	public T lire(int cle);
+	public T lire(Long cle);
 	public void inserer(T elt);
-	public void update(int index, T obj);
+	public void update(Long index, T obj);
 	public default void update(T obj) {}
-	public void effacer(int cle);
+	public void effacer(Long cle);
 }
